requirements.txt
azure-functions
openai
azure-core
azure-identity
python-dotenv